requirements.txt
requests
docker
six
pytest
pika
pyYaml